requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
click==8.1.7
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
Mako==1.3.8
MarkupSafe==3.0.2
numpy==2.2.0
openai==1.57.4
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.10.3
pydantic_core==2.27.1
PyMuPDF==1.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
